# perf: fetch abbreviated metadata