This simple change surprisingly improves performance by about 10%.

```sh
cargo run --bin=integrated-benchmark -- -s frozen-lockfile HEAD main -V -D ./crates/testing-utils/src/fixtures/big/
```

```
Benchmark 1: pacquet@HEAD
  Time (mean ± σ):      3.036 s ±  0.291 s    [User: 14.522 s, System: 17.387 s]
  Range (min … max):    2.673 s …  3.516 s    10 runs
 
Benchmark 2: pacquet@main
  Time (mean ± σ):      3.137 s ±  0.275 s    [User: 14.616 s, System: 17.753 s]
  Range (min … max):    2.584 s …  3.492 s    10 runs
 
Summary
  pacquet@HEAD ran
    1.03 ± 0.13 times faster than pacquet@main
```

It is still not as fast as pnpm. It is likely that there are other optimizations pnpm did that I'm not aware of.